+++ packages/core/src/messages/responses/generate-api-key.ts
